First pass at a pretty page for Captiv book recommendations
- [.gitignore]
Added ignore for the node_modules directory

- [Procfile]
File for heroku

- [app.js]
Configuration for the node app
	- Include the root routes
	- Serve /js and /css from the /public directory
	- Use Hogan for template rendering

- [app/routes/root.js]
Just render the page when the route is root. Waiting on parameters

- [app/templates/recommendedBook.html]
First pass at a page for book recommendations
Has a script tag so that we can render the correct book using variables passed from the route
	- Alternative would be to pull it out of the URL in javascript, but this seems like more work for no gain

- [app/templates/shared/header.html]
Shared file to import necessary javascript and css

- [config/env.js]
Basic ENVAR setting file

- [config/index.js]
Easier for importing the config files
Sets the APP_ROOT variable

- [package.json]
Defines dependencies

- [public/css/site.css]
CSS for the entire site

- [public/js/site.js]
Javascript for the whole site
Uses the book title and desired quote to serve up an embedded reader and an image of the cover. See comments for details (to be added…)